Adding Project Stubs for Race Engine Library
I'm going to need something to do the physics for the cars and stuff.
Since race car physics are not game dependent and the game is car
physics dependent, I'm making a separate project out of this. I'm also
going to need a way to store certain items, so the content project will
need to know how to read certain items, but not all items. This is why I
have a two project break down.