Race Track Editor: Saving Project File
Did some work on saving the project files. I made a small change to the
interface for the Project Model Writer. It no longer returns a bitmap
image when the image is set to a project, so the bitmap image should be
set to the project when within the implementations of the writer.