Race Track Editor: Reading & Writing Project File
I set the Uri under which I save the different project parts in the Zip
Reader Writer as readonly class level attributes. This insures there are
I don't have problems reading or writing any part even if there is a
typo in the Uri address string. I also got the application reading the
obstacle data file, so saving an opened project file won't cause the
obstacle data to be lost.