Race Track Editor: IO (Zip files)
The inherited model classes were causes issues for the XML Serialization
I was using to save the Obstacle Data Model. The way the Zip
implementation of core is shaping up, I don't really need those classes,
so I deleted them. The contents of the Obstacle Data Model are now now
both saved and read in the Zip implementation of Core.