Race Track Editor: Obstacle System
In the failed attempts to get the ObstacleLayerView to stop clipping the
content that would be rendered in the negative areas of the Canvas, I've
removed some code that wasn't really doing anything. Added a few more
test obstacles to use for development, and moved the ImageMargin back to
the race track image. When it some to exporting the obstacle data, I'll
just offset the obstacle positions by that margin. This will let me use
the top-left corner of the image as the point of origin (which has
always been the plan).