Race Track Editor: Obstacle System
Was working on dragging obstacles around the obstacle layer, but found
that the layer's measured canvas wasn't re-measuring its contents. Did a
few few things to force it re-measure; only found limited success this
way.  It worked more often then not, but there was a tendency to cut off
part of the items buy measuring just short of the actual size. I've set
this to have a min size that is 100 by 100 larger than the race track
images I'm using. This was hard coded into the xaml (a more dynamic
solution must be found).