Race Track Editor: Obstacle System
I've managed to get the Obstacle Editor to manipulate obstacles in any
Obstacle Layer as long as that layer is selected in the DataGrid on the
left side of the editor. Not quite the way I wanted it to work, but I
couldn't get overlapping Obstacle Layer Views contained with-in the same
ListBox to behave. I know that I can overlap multiple obstacle layer
views in code and not have a problem, which means my problem is with
either the list box or the list box item. Doesn't matter, using the list
box save on a lot of code I have to write.