Race Track Editor: Obstacle System
I was having an issue with the place holder item in the Data Grid
creating a Layer View with a context data item that wasn't of type
Obstacle Layer View Model. I switched to Collection View Source objects
to get rid of this issue. I was hoping this would let me interact with
obstacles in non-selected layers, but that part didn't pan out.