Physics
I created the convex polygon to convex polygon collision detection
method. It was identical to the triangle to convex polygon method
(except for the method signature). So, I pulled that out into it's own
method that both of those methods can call. I also found a small bug
that was introduced due to the use of the difference in position deltas
of the two bodies. Fixed it.